+++ lib/yokogaki.dart
@@ -5,6 +5,7 @@ library;
 // Widgets
 export 'src/widgets/horizontal_text.dart';
 export 'src/widgets/horizontal_rich_text.dart';
+export 'src/widgets/selectable_horizontal_text.dart';
 
 // Models
 export 'src/models/horizontal_text_style.dart';